--- components/hocs/indext.ts
@@ -1,5 +1,0 @@
-import withAuth from "./withAuth/withAuth";
-
-export {
-  withAuth,
-};
